rename collection and products dynamic routes
