+++ app/routes/$lang/collections/$collectionHandle.tsx
@@ -1,4 +1,4 @@
-import Component, { loader, meta } from '~/routes/collections/$collectionHandle';
+import Component, { loader, meta } from '~/routes/collections/$handle';
 
 export { loader, meta };
 export default Component;
