+++ app/routes/$lang/products/$productHandle.tsx
@@ -1,4 +1,4 @@
-import Component, { loader, action, ProductForm } from '~/routes/products/$productHandle';
+import Component, { loader, action, ProductForm } from '~/routes/products/$handle';
 
 export { loader, action, ProductForm };
 export default Component;
